refactor: Drizzle スキーマを認証とサンプル機能に分割する
