+++ src/lib/drizzle/auth-schema.ts
@@ -1,0 +1,80 @@
+import { sql } from "drizzle-orm";
+import { index, integer, sqliteTable, text, unique } from "drizzle-orm/sqlite-core";
+
+export const user = sqliteTable("user", {
+  createdAt: integer("created_at", { mode: "timestamp" })
+    .default(sql`(unixepoch())`)
+    .notNull(),
+  email: text("email").notNull().unique(),
+  emailVerified: integer("email_verified", { mode: "boolean" })
+    .$defaultFn(() => false)
+    .notNull(),
+  id: text("id").primaryKey(),
+  image: text("image"),
+  name: text("name").notNull(),
+  updatedAt: integer("updated_at", { mode: "timestamp" })
+    .default(sql`(unixepoch())`)
+    .notNull(),
+});
+
+export const session = sqliteTable(
+  "session",
+  {
+    createdAt: integer("created_at", { mode: "timestamp" }).default(sql`(unixepoch())`),
+    expiresAt: integer("expires_at", { mode: "timestamp" }).notNull(),
+    id: text("id").primaryKey(),
+    ipAddress: text("ip_address"),
+    token: text("token").notNull().unique(),
+    updatedAt: integer("updated_at", { mode: "timestamp" }).default(sql`(unixepoch())`),
+    userAgent: text("user_agent"),
+    userId: text("user_id")
+      .notNull()
+      .references(() => user.id, { onDelete: "cascade" }),
+  },
+  (table) => [
+    index("idx_session_expires_at").on(table.expiresAt),
+    index("idx_session_user_id").on(table.userId),
+  ],
+);
+
+export const account = sqliteTable(
+  "account",
+  {
+    accessToken: text("access_token"),
+    accessTokenExpiresAt: integer("access_token_expires_at", { mode: "timestamp" }),
+    accountId: text("account_id").notNull(),
+    createdAt: integer("created_at", { mode: "timestamp" }).default(sql`(unixepoch())`),
+    id: text("id").primaryKey(),
+    idToken: text("id_token"),
+    password: text("password"),
+    providerId: text("provider_id").notNull(),
+    refreshToken: text("refresh_token"),
+    refreshTokenExpiresAt: integer("refresh_token_expires_at", { mode: "timestamp" }),
+    scope: text("scope"),
+    updatedAt: integer("updated_at", { mode: "timestamp" }).default(sql`(unixepoch())`),
+    userId: text("user_id")
+      .notNull()
+      .references(() => user.id, { onDelete: "cascade" }),
+  },
+  (table) => [unique("idx_account_unique").on(table.providerId, table.accountId)],
+);
+
+export const verification = sqliteTable(
+  "verification",
+  {
+    createdAt: integer("created_at", { mode: "timestamp" }).default(sql`(unixepoch())`),
+    expiresAt: integer("expires_at", { mode: "timestamp" }).notNull(),
+    id: text("id").primaryKey(),
+    identifier: text("identifier").notNull(),
+    updatedAt: integer("updated_at", { mode: "timestamp" }).default(sql`(unixepoch())`),
+    value: text("value").notNull(),
+  },
+  (table) => [index("idx_verification_identifier").on(table.identifier)],
+);
+
+export const rateLimit = sqliteTable("rate_limit", {
+  count: integer("count").notNull(),
+  id: text("id").primaryKey(),
+  key: text("key").notNull().unique(),
+  lastRequest: integer("last_request").notNull(),
+});
